feat: restore full terminal width for alternate-screen apps (vim, etc.)
Detect alternate-screen transitions and resize the PTY to use the full
terminal width instead of subtracting the gutter margin. Also adds
default models.toml generation, clippy fixes, and idiomatic Rust
improvements across the codebase.

+++ crates/slate/src/ui/selection.rs
@@ -102,7 +102,7 @@ impl TextSelection {
                 continue;
             }
 
-            let vt_row = (screen_row - term_area_top) as u16;
+            let vt_row = screen_row - term_area_top;
 
             if vt_row >= screen_size.0 {
                 continue;
